JP-2707: Integrate move of comserver.event related components into already moved components (fixing import issues).

+++ comu/com.energyict.mdc.engine/src/main/java/com/energyict/mdc/engine/impl/web/events/commands/ComPortPoolRequestType.java
@@ -22,7 +22,7 @@ public class ComPortPoolRequestType extends IdBusinessObjectRequestType {
     }
 
     @Override
-    protected Request newRequestFor (Set<Integer> ids) {
+    protected Request newRequestFor (Set<Long> ids) {
         return new ComPortPoolRequest(ids);
     }
