+++ comu/com.energyict.mdc.engine/src/main/java/com/energyict/mdc/engine/impl/web/events/commands/ConnectionTaskRequestType.java
@@ -4,7 +4,7 @@ import java.util.Set;
 
 /**
  * Provides an implementation for the {@link RequestType} interface
- * for {@link com.energyict.mdc.tasks.ConnectionTask}s.
+ * for {@link com.energyict.mdc.device.data.tasks.ConnectionTask}s.
  *
  * @author Rudi Vankeirsbilck (rudi)
  * @since 2012-11-15 (16:59)
@@ -22,7 +22,7 @@ public class ConnectionTaskRequestType extends IdBusinessObjectRequestType {
     }
 
     @Override
-    protected Request newRequestFor (Set<Integer> ids) {
+    protected Request newRequestFor (Set<Long> ids) {
         return new ConnectionTaskRequest(ids);
     }
