+++ comu/com.energyict.mdc.engine/src/main/java/com/energyict/mdc/engine/impl/web/events/commands/DeviceRequestType.java
@@ -4,7 +4,7 @@ import java.util.Set;
 
 /**
  * Provides an implementation for the {@link RequestType} interface
- * for {@link com.energyict.mdc.protocol.api.device.BaseDevice devices}.
+ * for {@link com.energyict.mdc.device.data.Device}s.
  *
  * @author Rudi Vankeirsbilck (rudi)
  * @since 2012-11-15 (17:00)
@@ -22,7 +22,7 @@ public class DeviceRequestType extends IdBusinessObjectRequestType {
     }
 
     @Override
-    protected Request newRequestFor (Set<Integer> ids) {
+    protected Request newRequestFor (Set<Long> ids) {
         return new DeviceRequest(ids);
     }
